Replaced original Java SimpleDateFormat with IBM ICU implementation
This fixes sources which rely on the behavior of the Android implementation of SimpleDateFormat which is identical to the IBM ICU implementation.

The Java behavior of this class actually slightly differs from the Android version so nitpicky sources such as MangaSee will crash without this fix.

Replacing core Java libraries requires the JVM be booted with TachiServer in the bootstrap classpath however. Thus, TachiServer will now automatically boot a new JVM with itself in the bootstrap classpath if it detects the current JVM does not have itself in the bootstrap classpath.